[REFAC: #30] - Reorganizando os arquivos do backend.
close #30
