+++ src/backend/.env.sample
@@ -1,0 +1,1 @@
+TRIANG_INFOS_API_CORS="http://localhost:3000, https://localhost:3000"
