+++ src/backend/tests/api/conftest.py
@@ -1,0 +1,9 @@
+import pytest
+from fastapi.testclient import TestClient
+
+from api.app import app
+
+
+@pytest.fixture
+def client():
+    return TestClient(app)
